+++ __tests__/Observer.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test, vi } from "vitest";
-import { Observer } from '../src/index';
+import { Broadcaster, Observer } from '../src/index';
 
 describe('Observer', () => {
   test('can be constructed', () => {
@@ -12,5 +12,17 @@ describe('Observer', () => {
     const observer = new Observer(mockCallback);
     observer.onNotify({}, 'destroyed');
     expect(mockCallback).toHaveBeenCalledOnce();
+  });
+
+  test('can be subscribes to multiple Broadcasters', () => {
+    const mockCallback = vi.fn();
+    const observer = new Observer<Set<number>, 'borg' | 'test'>(mockCallback);
+    const broadcaster1 = new Broadcaster<Set<number>, 'test'>();
+    const broadcaster2 = new Broadcaster<Set<number>, 'borg'>();
+    broadcaster1.subscribe(observer);
+    broadcaster2.subscribe(observer);
+    broadcaster1.notify(new Set([5]), 'test');
+    broadcaster2.notify(new Set([5]), 'borg');
+    expect(mockCallback).toHaveBeenCalledTimes(2);
   });
 });