+++ src/broadcaster/Broadcaster.class.ts
@@ -4,9 +4,9 @@ import { BroadcastEvents } from "./BroadcastEvents.type";
 /**
  * Broadcaster to notify an {@link IObserver} of events.
  */
-export class Broadcaster<EventsType extends string> {
+export class Broadcaster<EntityType, EventsType extends string> {
   /** A set of {@link IObserver}s listening for events */
-  #listeners: Set<IObserver<EventsType>>;
+  #listeners: Set<IObserver<EntityType, EventsType>>;
 
   constructor() {
     this.#listeners = new Set();
@@ -16,7 +16,11 @@ export class Broadcaster<EventsType extends string> {
    * Subscribe an {@link IObserver} to notifications
    * @param observers the {@link IObserver}(s) to subscribe
    */
-  subscribe(observers: IObserver<EventsType> | IObserver<EventsType>[]) {
+  subscribe(
+    observers:
+      | IObserver<EntityType, EventsType>
+      | IObserver<EntityType, EventsType>[],
+  ) {
     if (observers instanceof Array) {
       for (const observer of observers) {
         this.#listeners.add(observer);
@@ -28,7 +32,7 @@ export class Broadcaster<EventsType extends string> {
    * Unsubscribe an {@link IObserver} from notifications
    * @param observer the {@link IObserver} to unsubscribe
    */
-  unsubscribe(observer: IObserver<EventsType>) {
+  unsubscribe(observer: IObserver<EntityType, EventsType>) {
     this.#listeners.delete(observer);
   }
 
@@ -37,7 +41,10 @@ export class Broadcaster<EventsType extends string> {
    * @param entity the entity which triggered the event
    * @param event the event type
    */
-  notify(entity: object, event: BroadcastEvents | EventsType) {
+  notify(
+    entity: EntityType | typeof this,
+    event: BroadcastEvents | EventsType,
+  ) {
     for (const listener of this.#listeners) {
       listener.onNotify(entity, event);
     }